zenpay_backend/api/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zenpay-backend"
version = "0.1.0"
description = "Backend for ZenPay usage-based billing service"
readme = "README.md"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "sqlalchemy>=1.4.0",
    "pydantic>=1.8.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.5",
    "stripe>=2.60.0",
    "alembic>=1.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "black>=21.5b0",
    "isort>=5.9.0",
    "mypy>=0.910",
    "flake8>=3.9.0",
]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?\$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true